skills/ai-product-analyst-main/data-generation/requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
duckdb>=0.9
faker>=20.0
tqdm>=4.65
